% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_fun_compare_time.R
\name{algorithm_speed}
\alias{algorithm_speed}
\title{Title:  Function to run a list of algorithm multiple times and check the time it takes them to sort the same data}
\usage{
algorithm_speed(algorithm = list(bubble, quick), argument, repetition)
}
\arguments{
\item{algorithm}{The sorting algorithms you wish to compare. Present them in a list. As a default it will be list(bubble_sort, quick_sort)}

\item{argument}{The argument that needs to be sorted. It can take both numeric values and characters}

\item{repetition}{The number of time the sorting is performed.}
}
\value{
The function will return 2 lists of times (or more if you made more sorting algorithms to compare),
where the times listed is the time it took to sort the  argument at each repetition. The function also print the list of names
of the algorithm input. Now you can compare the time it took to sort the chosen argument.
}
\description{
The \emph{algorithm_speed()} compares the speed of different sorting algorithms.
The package \emph{Wind} provide the function for a bubble sort (bubble)
and a quick sort (quick). Thus, those can be used as inputs. However, if one wish to
compare speed of other sorting algorithms, one can write those and use them as input
in \emph{algorithm_speed()}.
}
\examples{
# Examples of different function call
algorithm_speed(algorithm = list(quick, bubble), argument = sample(1:1000), repetition = 10)
algorithm_speed(argument = sample(1:1000), repetition = 10)
algorithm_speed(argument = LETTERS, repetition = 10) # it can also take letters
}
