% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_fun_compare_time.R
\name{check_algorithm}
\alias{check_algorithm}
\title{Title:  Function to run algorithm multiple times and check the time it takes
Argument algorithm: the package provide the function for a bubble sort (bubble_sort)
and a quick sort (quick_sort). Thus, those can be used as inputs. If one wish to
compare speed of other functions, write those and use them as input.}
\usage{
check_algorithm(algorithm = list(bubble_sort, quick_sort), arr, number)
}
\arguments{
\item{algorithm}{The sorting algortihms you wish to compare. Present them in a list. As a default it will be list(bubble_sort, quick_sort)}

\item{arr}{The argument that needs to be sorted. It can take both numeric values and characters}

\item{number}{The number of time the sorting is compared}
}
\value{
Time it takes to sort the argument
}
\description{
Title:  Function to run algorithm multiple times and check the time it takes
Argument algorithm: the package provide the function for a bubble sort (bubble_sort)
and a quick sort (quick_sort). Thus, those can be used as inputs. If one wish to
compare speed of other functions, write those and use them as input.
}
\examples{
Function call
check_algorithm(algorithm = list(quick_sort, bubble_sort), arr = sample(1:1000), number = 10)
check_algorithm(arr = sample(1:1000), number = 10)
check_algorithm(arr = LETTERS, number = 10) # it can also take letters
 example data
df <- c("e","t","p","y")
check_algorithm(arr = df, number = 10)
}
