% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_fun_sliding_window.R
\name{sliding_window}
\alias{sliding_window}
\title{Title: Sliding window function}
\usage{
sliding_window(data, window_size, window_type)
}
\arguments{
\item{data}{The sequential input data}

\item{window_size}{The size of the window}

\item{window_type}{The window type. window_type takes one of 2 inputs:
"Max_diff" or "MinVal". This will make the function run either window_Max_diff
or window_Min_val, respectively}
}
\value{
Orginal data with an extra column
}
\description{
Title: Sliding window function
}
\examples{
Calling the function with: window size 10 and a threshold = 5.
sliding_window(Test_data_frame, 10, "Max_diff")
}
